Factor out an mbean cache
